wasm2js: Add a `--only-asmjs-wrapper` argument
This commit adds a new argument, `--only-asmjs-wrapper`, to the `wasm2js` tool.
While probably not going to be widely used it's intended for this argument to be
used by the test suite to continue operating as it does today. When not passed
the behavior of the `wasm2js` tool will be to emit an ES module by default (to
be added after this commit).