Improve getFallthrough (#1643)
That method looks through tee_locals and other operations that receive a value and let it flow through them, like a block's final value, etc. It just handled a few such operations, with this PR all of them should be handled.

Also refactor it out of the OptimizeInstructions pass as I think it may be useful for propagating returned constants.