[wasm2asm] Add --compact-function-tables option that eliminate duplicates based on emcc output.
Right now it's not possible to compile large projects that require many function pointers and ALIASING_FUNCTION_POINTERS=0.
The resulting function table is too big to be loadable by browsers.

This change exploits the structure of the asm.js output produced by emcc to compact the function table and
eliminate all padding functions.