DeadArgumentElimination Pass (#1641)
This adds a pass to remove unnecessary call arguments in an LTO-like manner, that is:

*    If a parameter is not actually used in a function, we don't need to send anything, and can remove it from the function's declaration. Concretely,

(func $a (param $x i32)
 ..no uses of $x..
)
(func $b
  (call $a (..))
)
=>
(func $a
 ..no uses of $x..
)
(func $b
  (call $a)
)

And

*    If a parameter is only ever sent the same constant value, we can just set that constant value in the function (which then means that the values sent from the outside are no longer used, as in the previous point). Concretely,

(func $a (param $x i32)
  ..may use $x..
)
(func $b
  (call $a (i32.const 1))
  (call $a (i32.const 1))
)
=>
(func $a
  (local $x i32)
  (set_local $x (i32.const 1)
  ..may use $x..
)
(func $b
  (call $a)
  (call $a)
)

How much this helps depends on the codebase obviously, but sometimes it is pretty useful. For example, it shrinks 0.72% on Unity and 0.37% on Mono. Note that those numbers include not just the optimization itself, but the other optimizations it then enables - in particular the second point from earlier leads to inlining a constant value, which often allows constant propagation, and also removing parameters may enable more duplicate function elimination, etc. - which explains how this can shrink Unity by almost 1%.

Implementation is pretty straightforward, but there is some work to make the heavy part of the pass parallel, and a bunch of corner cases to avoid (can't change a function that is exported or in the table, etc.). Like the Inlining pass, there is both a standard and an "optimizing" version of this pass - the latter also optimizes the functions it changes, as like Inlining, it's useful to not need to re-run all function optimizations on the whole module.