Fix segment size validation for imported memories
Without this wasm-opt can't operation on emscripten-produced
SIDE_MODULES's which have zero sized memory imports.

Technically is not a validation failure if you have segments that
are larger than your initial memory, regardless of whether you
import them.

For non-imported memories it can be helpful though, so leaving it
in to catch those errors.