Merge-Blocks improvements (#1760)
Previously we didn't try to merge a block into the parent if the block had a name. This lets us merge part of it, that is:

(block
  (..a..)
  (block $child
    (..b..)
    (.. some br to $child ..)
    (..c..)
  )
)
=>
(block
  (..a..)
  (..b..) ;; moved out
  (block $child
    (.. some br to $child ..)
    (..c..)
  )
)
This is beneficial for 2 reasons: the child may now be a singleton, so we can remove the block; or, now that we canonicalized the br-containing code to the head of the child, we may be able to turn it into an if.