LocalCSE fuzz fix: invalidate the set operations too (#1778)
We invalidated based on effects of set values, but not of the sets themselves. Without that, a set could be overridden by something irrelevant and we thought we could still reuse the old value.

Before this PR, the testcase would have the last set's value be optimized into a get, incorrectly.