wasm-emscripten-finalize: Remove stack pointer global from shared libs (#1791)
The wasm backend uses a wasm global (__stack_pointer) for the shadow
stack location.   In order to make this work with shared libraries the
main module would have to export this global and shared libraries would
need to import it.  This means we'd be relying of mutable globals which
are not yet implemented in all browsers.

This change allows is to move forward with shared libraries without
mutable global support by replacing all stack pointer access in shared
libraries with functions calls.