Move if copy logic from coalesce-locals to remove-unused-brs.
If copies is the case where an if arm is a get that feeds into
a set of the same local:

  (set_local $x
    (if (result i32)
      (..condition..)
      (..result)
      (get_local $x)
    )
  )

We can rework this so that the if-else is only an if, which
executes the code path not going to the get.

This was done in coalesce-locals only because it is likely to
work there as after coalescing there are more copies. However,
the logic is of removing a branch, and so belongs in
remove-unused-brs, and fits alongside existing logic there
for handling ifs with an arm that is a br. Also refactor that
code so that the two optimizations can feed into each other.