Speculate in simplify-locals that it is worth turning an if into an if-else.
If an if sets a local,

  (if
    (..condition..)
    (set_local $x (..value..))
  )

we can turn it into

  (set_local $x
    (if
      (..condition..)
      (..value..)
      (get_local $x)
    )
  )

This increases code size and adds a branch in the if, but allows
the set to be optimized into a tee or optimized out entirely. In
the worst case, other optimizations can break up an if with a copy
in one of its arms later.

Includes a determinism fix for EquivalentSets, which this patch
triggered.