Simplify ExpressionAnalyzer (#1920)
This refactors the hashing and comparison code to use a single immediate-value iterator. This makes us have a single place that knows the list of immediate fields in every node type, instead of 2.

This also fixes a few bugs found by doing that. In particular, this makes us slightly slower than before since we are hashing more fields.