Fix OSX build error on master
Accidentally introduced in #1698, it looks like the
`MACOSX_DEPLOYMENT_TARGET` can't be set so low to 10.7 because it causes
build errors. After some [investigation] it looks like the minimum
currently required is 10.9, so this commit changes the OSX deployment
target to 10.9 in an effort to be compatible with as many OSX
installations as possible.

[investigation]: https://travis-ci.org/WebAssembly/binaryen/builds/509462566