Change default feature set to MVP
This greatly simplifies the feature handling logic and does not
change the user experience when a target features section is present.
Also refactors features to be stored on the Module after they are read
out of the target features section, allowing us to emit a (potentially
modified) target features section and conditionally emit sections
based on features, such as the DataCount section.