Fuzz fixes (#1991)
Get fuzzer to attempt to create almost all features. Pass v8 all the flags to allow that.

Fix fuzz bugs where we read signed_ even when it was irrelevant for that type of load.

Improve wasm-reduce on fuzz testcases, try to replace a node with drops of its children, not just the children themselves.