Change default feature set to MVP (#1993)
In the absence of the target features section or command line flags. When there are command line flags, it is an error if they do not exactly match the target features section, except if --detect-features has been provided.

Also adds a --print-features pass to print the command line flags for all enabled options and uses it to make the feature tests more rigorous.