Parsing, validation, and interpretation of Push/Pop
DO NOT USE THESE INSTRUCTIONS YET.

Adds Push and Pop instructions to explicitly manipulate the value
stack. Implements interpretation, s-expression parsing, and
validation. The parser is now required to pass a depth parameter
around because the type of a Pop instruction is context-dependent.

Also implements an initial, untested new side effect for stacky code
that should make pushes and pops opaque to the optimizer. A follow-on
PR will implement fuzzer support for the new instructions and add
regression tests for any bugs it uncovers.

A later follow-on PR will use these instructions to implement support
for multivalue functions and calls.

Next the binary reader will be replaced with a new implementation that
parses binaries into a completely flat, stacky structure that is
folded by a subsequent pass. This will gracefully allow for multivalue
blocks and any other multivalue instructions.

Finally Stack IR optimizations will be replaced with optimizations on
the flat form of Binaryen IR and Stack IR will be deleted.