First stage of cleeanup in source tree pollution (#2105)
Update build-js.sh to output to `out` directory.  This is district
from the `bin` directory which is used by the cmake build and may or
may not live in the source tree.  The `out` directory currently always
lives in the source tree.

As a followup change I hope to additionally move all test outout into
this tree.

See #2104