Make clang-tidy-diff.sh run on background
Currently if a command in Travis CI exceeds 10 mins it fails with the
message `No output has been received in the last 10m0s`. But
clang-tidy-diff.sh might take longer than that if the number of modified
files is large.

According to
https://docs.travis-ci.com/user/common-build-problems/#Build-times-out-because-no-output-was-received,
prefixing the command with `travis mvn` spawns a new process running the
command. `travis_wait` then writes a short line to the build log every
minute for 20 minutes, extending the amount of time your command has to
finish.