Split clag-tidy-diff.sh into two parts
Currently clang-tidy-diff.sh runs clang-tidy twice, because clang-tidy
does not return nonzero exit code when there are warnings. So currently
clang-tidy-diff.sh:
1. Runs clang-tidy and redirect its stdout to a variable `TIDY_MSG`, and
   check if the `TIDY_MSG` does not contain any warnings
2. If `TIDY_MSG` contains warnings, runs clang-tidy again to print the
   error to the console (echo'ing `TIDY_MSG` does not achieve the same
   thing. It scrambles messages and delete all colors), and exits with a
   nonzero code.

But clang-tidy can take a long time when the number of modified files
is large, and running it twice sometimes can exceeds 10 mins, which is
the limit of Travis CI command.

So we split the script into two parts here.
1. Runs clang-tidy and show whatever it prints.
2. *In a separate script*, runs it again and redirects its output to
   `TIDY_MSG`, and if it contains any warnings, exits with a nonzero
   code.

Also this moves clang-tidy after all other tests, so now even if it
fails with a very large number of changed files, which is unlikely with
the split scripts anyway, we will see if all the other tests succeed
before that.