Bysyncify: optimize better by coalescing before instrumenting control flow (#2183)
This results in better code sizes on many testcases, sometimes much better. For example, on SQLite the 150K function has only 27 locals instead of 3,874 which it had before (!). This also reduces total code size on SQLite by 15%.

The key issue is that after instrumenting control flow we have a lot bigger live ranges.

This must be done rather carefully, as we need to introduce some temp locals early on (for breaking up ifs, for call return values, etc.).