Ignore --initial-stack-pointer arg to wasm-emscripten-finalize
We were passing bad value in --initial-stack-pointer which did not
include the STATIC_BUMP (since STATIC_BUMP is determinted by the output
of finalize).

If emscripten wants to set the stack pointer position it can do
so by calling the stackRestore() function at startup.

This argument will be removed completely once we stop passing it on the
emscripten side.

See https://github.com/emscripten-core/emscripten/issues/8905