wasm-emscripten-finalize: Internalize mutable __stack_pointer import
I'm working on a change to lld that will cause `-pie` binaries to
import __stack_pointer, just like -shared do already.  Because we
don't yet support mutable globals everywhere this change will
internalize the import and create a new immutable import that is used
to initialize the internal one.

This change is part of the fix for:
https://github.com/emscripten-core/emscripten/issues/8915