wasm-emscripten-finalize: Add mainReadsParams metadata (#2247)
The new flag indicates whether main reads the argc/argv parameters. If it does not, we can avoid emitting code to generate those arguments in the JS, which is not trivial in small programs - it requires some string conversion code.

Nicely the existing test inputs were enough for testing this (see outputs).

This depends on an emscripten change to land first, as emscripten.py asserts on metadata fields it doesn't recognize.